+++ src/constants.ts
@@ -89,6 +89,7 @@ export const enum Commands {
     CreateNewJql = 'atlascode.jira.createNewJql',
     ToDoIssue = 'atlascode.jira.todoIssue',
     RovodevOpenHelp = 'atlascode.rovodev.openHelp',
+    RovodevEnable = 'atlascode.rovodev.enable',
     InProgressIssue = 'atlascode.jira.inProgressIssue',
     DoneIssue = 'atlascode.jira.doneIssue',
     ShowOnboardingFlow = 'atlascode.showOnboardingFlow',